+++ GW2EIEvtcParser/EIData/ProfHelpers/Thief/SpecterHelper.cs
@@ -43,7 +43,7 @@ namespace GW2EIEvtcParser.EIData
             new Buff("Consume Shadows", ConsumeShadows, Source.Specter, ArcDPSEnums.BuffStackType.StackingConditionalLoss, 5, BuffClassification.Other, BuffImages.ConsumeShadows),
         };
 
-        private static HashSet<long> Minions = new HashSet<long>()
+        private static HashSet<int> Minions = new HashSet<int>()
         {
             (int)MinionID.Specter1,
             (int)MinionID.Specter2,
@@ -57,7 +57,7 @@ namespace GW2EIEvtcParser.EIData
             (int)MinionID.Specter10,
         };
 
-        internal static bool IsKnownMinionID(long id)
+        internal static bool IsKnownMinionID(int id)
         {
             return Minions.Contains(id);
         }
